#22 changed aruco projections of pixels to float. Optimizer seems to work much better.

+++ OptimizationUtils/utilities.py
@@ -252,7 +252,7 @@ def projectToCamera(intrinsic_matrix, distortion, width, height, pts):
 
     # Project the 3D points in the camera's frame to image pixels
     # From https://docs.opencv.org/2.4/modules/calib3d/doc/camera_calibration_and_3d_reconstruction.html
-    pixs = np.zeros((2, n_pts), dtype=np.int)
+    pixs = np.zeros((2, n_pts), dtype=np.float)
 
     k1, k2, p1, p2, k3 = distortion
     # fx, _, cx, _, fy, cy, _, _, _ = intrinsic_matrix
